CHATBOT-381 - Fix bot deletion

+++ vertigo-bot-designer/src/main/java/io/vertigo/chatbot/designer/builder/services/bot/ChabotCustomConfigServices.java
@@ -45,7 +45,9 @@ public class ChabotCustomConfigServices implements Component {
 
 	public void deleteChatbotCustomConfig(@SecuredOperation("botAdm") final Chatbot bot) {
 		ChatbotCustomConfig chatbotCustomConfig = getChatbotCustomConfigByBotId(bot.getBotId());
-		chatbotCustomConfigDAO.delete(chatbotCustomConfig.getCccId());
+		if (chatbotCustomConfig.getCccId() != null) {
+			chatbotCustomConfigDAO.delete(chatbotCustomConfig.getCccId());
+		}
 	}
 
 	public ChatbotCustomConfig getChatbotCustomConfigByBotId(final Long botId) {
